add mail to admins

+++ database/migrations/2024_01_25_220239_alter_table_kits_cllection_group_id_to_nullable_migration.php
@@ -32,7 +32,9 @@ return new class extends Migration
     {
         Schema::table('kits', function (Blueprint $table) {
             //
-            $table->dropColumn('collection_group_id');
+            if (Schema::hasColumn('kits', 'collection_group_id')) {
+                $table->dropColumn('collection_group_id');
+            }
         });
         Schema::table('kits', function (Blueprint $table) {
             //
